requirements.txt
# AI System Administrator Agent Requirements
# Core AI Framework
pyautogen>=0.2.0
openai>=1.0.0

# LLM Backend
llama-cpp-python>=0.2.0
huggingface-hub>=0.20.0

# Web Interface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0

# CLI Interface
click>=8.1.0
rich>=13.0.0
prompt-toolkit>=3.0.0

# System Integration
psutil>=5.9.0
subprocess32>=3.5.4; python_version < "3.3"

# Security & Logging
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
structlog>=23.0.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Docker support
docker>=6.1.0
